requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface[community]
langchain-openai
langchain-groq
langchain-google-genai
langchain-mistralai
langchain-huggingface
streamlit
langchain_openrouter
langchain-tavily
langgraph